+++ lib/src/main/java/com/wjduquette/joe/Scanner.java
@@ -84,11 +84,19 @@ class Scanner {
             case ':'  -> addToken(COLON);
             case ','  -> addToken(COMMA);
             case '.'  -> addToken(DOT);
-            case '-'  -> addToken(match('>') ? MINUS_GREATER : MINUS);
-            case '+'  -> addToken(PLUS);
+            case '-'  -> {
+                if (match('=')) {
+                    addToken(MINUS_EQUAL);
+                } else if (match('>')) {
+                    addToken(MINUS_GREATER);
+                } else {
+                    addToken(MINUS);
+                }
+            }
+            case '+'  -> addToken(match('=') ? PLUS_EQUAL : PLUS);
             case '?'  -> addToken(QUESTION);
             case ';'  -> addToken(SEMICOLON);
-            case '*'  -> addToken(STAR);
+            case '*'  -> addToken(match('=') ? STAR_EQUAL : STAR);
             case '!'  -> addToken(match('=') ? BANG_EQUAL : BANG);
             case '='  -> addToken(match('=') ? EQUAL_EQUAL : EQUAL);
             case '<'  -> addToken(match('=') ? LESS_EQUAL : LESS);
@@ -97,6 +105,8 @@ class Scanner {
                 if (match('/')) {
                     // A comment goes until the end of the line.
                     while (peek() != '\n' && !isAtEnd()) advance();
+                } else if (match('=')) {
+                    addToken(SLASH_EQUAL);
                 } else {
                     addToken(SLASH);
                 }